Effectively final version of dapc tutorial.
I cut everything out about PCA because DAPC is pretty easy and uses a PCA in the analysis
I tried to annotate everything well enough for a new user to understand.

Chose k=2 for DAPC based on paper's k-clustering, as decided by a more appropriate method external to this tutorial.

Plotted one plot at the end to show distances between clusters.

Included a brief line on identifying which sample belongs to each group but did not fully demonstrate building a convenient table.