Adding a script to run Admixture.
This script will run the maximum-likelihood population structure
estimation software Admixture on an input file provided by the user on
the command line. Using a for loop, this iterates over a variety of "K"
values, creating Admixture outputs to be compared for the identification
of the optimal number of populations.